Added some commands, better ErrorLogging, and JSON settings file!
JSON Config and Custom commands file can be found at the root folder.
Paste them into /Essentials/

+++ BP-Essentials/BP-Essentials/Chat/Commands/Everyone/ClearChat.cs
@@ -11,19 +11,19 @@ namespace BP_Essentials.Commands {
                 string arg1 = GetArgument.Run(1, false, false, message).Trim();
                 if (arg1 == "all" || arg1 == "everyone") {
                     if (AdminsListPlayers.Contains(player.playerData.username)) {
-                        player.SendToAll(Channel.Unsequenced, (byte) 10, "Clearing chat for everyone...");
+                        player.SendToAll(Channel.Unsequenced,  10, "Clearing chat for everyone...");
                         Thread.Sleep(500);
                         for (var i = 0; i < 6; i++)
-                            player.SendToAll(Channel.Unsequenced, (byte) 10, " ");
+                            player.SendToAll(Channel.Unsequenced,  10, " ");
                     }
                     else
-                        player.SendToSelf(Channel.Unsequenced, (byte) 10, MsgNoPerm);
+                        player.SendToSelf(Channel.Unsequenced,  10, MsgNoPerm);
                 }
                 else {
-                    player.SendToSelf(Channel.Unsequenced, (byte) 10, "Clearing the chat for yourself...");
+                    player.SendToSelf(Channel.Unsequenced,  10, "Clearing the chat for yourself...");
                     Thread.Sleep(500);
                     for (var i = 0; i < 6; i++)
-                        player.SendToSelf(Channel.Unsequenced, (byte) 10, " ");
+                        player.SendToSelf(Channel.Unsequenced,  10, " ");
                 }
             }
             catch (Exception ex) {
